Fix modals
Since roadmap overview was moved higher level router, roadmap deletion
modals did not open. Moved modal opening to MainRouter.

Also fixed EditVersionModal's styles.
