+++ frontend/src/components/modals/EditVersionModal.module.scss
@@ -11,6 +11,6 @@
   }
 }
 
-.formDiv {
+.versionformDiv {
   margin-bottom: 16px;
 }
